Move QueryType to presto-common
