+++ presto-common/src/main/java/com/facebook/presto/common/resourceGroups/QueryType.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.spi.resourceGroups;
+package com.facebook.presto.common.resourceGroups;
 
 import com.facebook.drift.annotations.ThriftEnum;
 import com.facebook.drift.annotations.ThriftEnumValue;
